move quests actions into its own router
