+++ packages/backend/src/handlers/actions/quests/routes.ts
@@ -1,0 +1,8 @@
+import express from 'express';
+
+import { asyncHandlerWrapper } from '../../../lib/apiHelpers';
+import { createQuestHandler } from './createQuest/handler';
+
+export const questsRoutes = express.Router();
+
+questsRoutes.post('/create', asyncHandlerWrapper(createQuestHandler));
